some code shuffle, move mutex to inner funcs
